Fix state being drawn and destroyed at the same time
On slow phones it is very likely for the state to take too long
to be destroyed and for the graphics thread to try to draw the
state object in the middle of the destroy process.
This is fixed by making the global state pointer NULL before
calling delete on the state object, thereby allowing the graphics
thread to see that there is no valid state object.